requirements.txt
apiclient==1.0.4
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2018.1.18
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
discovery==0.0.4
distlib==0.3.1
Django==1.10.6
docutils==0.13.1
filelock==3.0.12
future==0.18.2
#google==3.0.0
#google-api-core==1.22.4
#google-api-python-client==1.12.3
#google-auth==1.22.1
#google-auth-httplib2==0.0.4
#google-auth-oauthlib==0.4.1
#googleapis-common-protos==1.52.0
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.18.1
idna==2.6
oauth2client==4.1.3
oauthlib==3.1.0
packaging==16.8
pip==20.2.4
pipenv==11.8.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.2.0
pyowm==2.8.0
pyparsing==2.2.0
PySocks==1.7.1
pyTelegramBotAPI==3.6.1
python-dateutil==1.5
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
